\name{createSimulationTable}
\title{Generic function}
\usage{createSimulationTable(x, ...)}
\description{Generic function}
\alias{createSimulationTable}
\author{TszKin Julian Chan \email{ctszkin@gmail.com}}
\keyword{internal}
\seealso{\code{\link{createSimulationTable.ezsim}}}
\arguments{\item{x}{Object}
\item{\dots}{Further arguments}}
