% Generated by roxygen2 (4.0.1): do not edit by hand
\name{generate}
\alias{generate}
\title{Generic function}
\usage{
generate(x, ...)
}
\arguments{
\item{x}{Object}

\item{\dots}{Further arguments}
}
\description{
Generic function
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}
\seealso{
\code{\link{generate.parameterDef}}
}
\keyword{internal}

