\name{getScalarsName}
\title{Generic function}
\usage{getScalarsName(x, ...)}
\description{Generic function}
\alias{getScalarsName}
\author{TszKin Julian Chan \email{ctszkin@gmail.com}}
\seealso{\code{\link{getScalarsName.ezsim}}, \code{\link{getScalarsName.summary.ezsim}}}
\arguments{\item{x}{Object}
\item{\dots}{Further arguments}}
