% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getSelectionName}
\alias{getSelectionName}
\title{Generic function}
\usage{
getSelectionName(x, ...)
}
\arguments{
\item{x}{Object}

\item{\dots}{Further arguments}
}
\description{
Generic function
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}
\seealso{
\code{\link{getSelectionName.ezsim}}, \code{\link{getSelectionName.summary.ezsim}}
}
\keyword{internal}

