\name{label_both_parsed_recode}
\title{Combine label_both and label_parsed in \pkg{ggplot2}.}
\usage{label_both_parsed_recode(display_name)}
\description{Combine label_both and label_parsed in \pkg{ggplot2}. Also added a rename function to it
see label_both and label_parsed in \pkg{ggplot2} for details.}
\alias{label_both_parsed_recode}
\value{A function similar to label_both and label_parsed in \pkg{ggplot2} for details.}
\author{TszKin Julian Chan \email{ctszkin@gmail.com}}
\references{\url{http://cran.r-project.org/web/packages/ggplot2/index.html}}
\seealso{\code{\link{label_both}}, \code{\link{label_parsed}}}
\keyword{internal}
\arguments{\item{display_name}{A vector contains the display name. Names of the vector are the original name.}}
