\name{print.ezsim}
\title{Print an ezsim Object.}
\usage{\method{print}{ezsim} (x, ...)}
\description{Print an ezsim Object. See \code{\link{ezsim}} for details.}
\alias{print.ezsim}
\author{TszKin Julian Chan \email{ctszkin@gmail.com}}
\seealso{\code{\link{ezsim}}}
\arguments{\item{x}{An ezsim Object}
\item{\dots}{unused}}
