\name{print.parameterDef}
\title{Print a parameterDef Object.}
\usage{\method{print}{parameterDef} (x, ...)}
\description{Print a parameterDef Object in the console}
\alias{print.parameterDef}
\author{TszKin Julian Chan \email{ctszkin@gmail.com}}
\seealso{\code{\link{createParDef}}}
\arguments{\item{x}{A parameterDef Object}
\item{\dots}{unused}}
