\name{print.parameters}
\title{Print a parameters Object.}
\usage{\method{print}{parameters} (x, ...)}
\description{Print a parameters Object in the console}
\alias{print.parameters}
\author{TszKin Julian Chan \email{ctszkin@gmail.com}}
\seealso{\code{\link{parameterDef}}}
\arguments{\item{x}{A parameters Object}
\item{\dots}{unused}}
