\name{print.summary.ezsim}
\title{Print a summary.ezsim Object.}
\usage{\method{print}{summary.ezsim}(x,digits=4,...)}
\description{Print a summary.ezsim Object in the console. See \code{\link{summary.ezsim}} for details}
\alias{print.summary.ezsim}
\author{TszKin Julian Chan \email{ctszkin@gmail.com}}
\seealso{\code{\link{summary.ezsim}}}
\arguments{\item{x}{A summary.ezsim Object}
\item{digits}{Number of digits the data will be rounded to.}
\item{\dots}{unused}}
