\name{run}
\alias{run}
\title{Generic function}
\usage{
  run(x, ...)
}
\arguments{
  \item{x}{Object}

  \item{\dots}{Further arguments}
}
\description{
  Generic function
}
\author{
  TszKin Julian Chan \email{ctszkin@gmail.com}
}
\seealso{
  \code{\link{run.ezsim}}
}
\keyword{internal}

