% Generated by roxygen2 (4.0.1): do not edit by hand
\name{test}
\alias{test}
\title{Generic function}
\usage{
test(x, ...)
}
\arguments{
\item{x}{Object}

\item{\dots}{Further arguments}
}
\description{
Generic function
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}
\seealso{
\code{\link{test.ezsim}}
}
\keyword{internal}

