\name{createSimulationTable.ezsim}
\title{createSimulationTable}
\usage{createSimulationTable.ezsim(x)}
\description{Create a simulation table from simulation result}
\alias{createSimulationTable}
\alias{createSimulationTable.ezsim}
\value{an ezsim object}
\author{TszKin Julian Chan \email{ctszkin@gmail.com}}
\arguments{\item{an}{ezsim object}}
\examples{\dontrun{

}}
