% Generated by roxygen2 (4.0.1): do not edit by hand
\name{merge.ezsim}
\alias{merge.ezsim}
\title{Merge two ezsim objects}
\usage{
\method{merge}{ezsim}(x, y, ...)
}
\arguments{
\item{x}{A ezsim to merge with}

\item{y}{A ezsim to merge with}

\item{...}{unused}
}
\description{
Merge two ezsim objects. Either \code{m} or \code{parameter_def} of two ezsim objects must be the same. If \code{parameter_def} are the same, the merging is regarded as increasing the number of simulation \code{m}. If the \code{parameter_def} are different and \code{m} are the same, the merging is regarded as extending the \code{parameter_def}.
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}
\seealso{
\code{\link{ezsim}}
}

