% Generated by roxygen2 (4.0.1): do not edit by hand
\name{merge.parameterDef}
\alias{merge.parameterDef}
\title{Merge two parameterDef objects}
\usage{
\method{merge}{parameterDef}(x, y)
}
\arguments{
\item{x}{A parameterDef to merge with}

\item{y}{A parameterDef to merge with}
}
\description{
Merge two parameterDef objects. 'others' of two parameterDef objects must be the same. 'scalars' of two parameterDef objects must have same name and the value must not overlap.
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}
\seealso{
\code{\link{createParDef}}, \code{\link{createParDef}}
}

