% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.ezsim}
\alias{print.ezsim}
\title{Print an ezsim Object.}
\usage{
\method{print}{ezsim}(x, ...)
}
\arguments{
\item{x}{An ezsim Object}

\item{\dots}{unused}
}
\description{
Print an ezsim Object. See \code{\link{ezsim}} for details.
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}
\seealso{
\code{\link{ezsim}}
}

