% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.parameters}
\alias{print.parameters}
\title{Print a parameters Object.}
\usage{
\method{print}{parameters}(x, ...)
}
\arguments{
\item{x}{A parameters Object}

\item{\dots}{unused}
}
\description{
Print a parameters Object in the console
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}
\seealso{
\code{\link{parameterDef}}
}

