\name{setSelection}
\title{Generic function}
\usage{setSelection(x, ...)}
\description{Generic function}
\alias{setSelection}
\author{TszKin Julian Chan \email{ctszkin@gmail.com}}
\keyword{internal}
\seealso{\code{\link{setSelection.parameterDef}}}
\arguments{\item{x}{Object}
\item{\dots}{Further arguments}}
