% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setSelection}
\alias{setSelection}
\title{Generic function}
\usage{
setSelection(x, ...)
}
\arguments{
\item{x}{Object}

\item{\dots}{Further arguments}
}
\description{
Generic function
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}
\seealso{
\code{\link{setSelection.parameterDef}}
}
\keyword{internal}

